backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-generativeai==0.8.4
qdrant-client==1.8.0
asyncpg==0.29.0
python-dotenv==1.0.0
aiofiles==23.2.1